// src/modules/admin/map/types.ts
// Department Listing Table
export interface DepartmentRow {
	department_id: string;
	department_name: string;
	department_code: string;
	dept_status: string;
}

// Zone Listing Table
export interface ZoneRow {
	zone_id: string;
	zone_name: string;
	zone_code: string;
	zone_status: string;
}
// Branch Listing Table
export interface BranchRow {
	branch_id: string;
	branch_name: string;
	branch_code: string;
	branch_status: string;
}